% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{cn}
\alias{cn}
\title{Calculate CN site-dependent profile}
\usage{
cn(prot)
}
\arguments{
\item{prot}{is a protein object obtained using enm()}
}
\value{
a vector of size nsites with cn values for each site
}
\description{
Calculates the Contact Number (CN) of each site
}
\seealso{
Other site profiles: 
\code{\link{dactive}()},
\code{\link{msfi}()},
\code{\link{wcn}()}
}
\concept{site profiles}
