% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{dactive}
\alias{dactive}
\title{Distance to Active Site}
\usage{
dactive(prot, pdb_site_active)
}
\arguments{
\item{prot}{A protein object created by [enm()]}

\item{pdb_site_active}{Vector of PDB residue numbers defining the active site}
}
\value{
a vector of size nsites with dactive values for each site
}
\description{
Calculates the distance from each site to the closest active site residue.
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{dg_ent_act}()},
\code{\link{dv_act}()},
\code{\link{g_ent}()},
\code{\link{msfi}()},
\code{\link{msfn}()},
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
