% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Ddg_ent_act}
\alias{Ddg_ent_act}
\title{Change in Activation Energy (Entropic)}
\usage{
Ddg_ent_act(wt, mut, ideal = wt, pdb_site_active = NA, beta = beta_boltzmann())
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}

\item{ideal}{Reference protein for ideal active site conformation}

\item{pdb_site_active}{Vector of PDB residue numbers defining the active site}

\item{beta}{Inverse temperature (1/kT). Default: beta_boltzmann()}
}
\value{
a scalar energy difference (NA if pdb_site_active is NA)
}
\description{
Calculates the change in entropic contribution to activation energy between
wild-type and mutant.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Ddv_act}()},
\code{\link{Dg_ent}()},
\code{\link{Dmsfi}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
