% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Ddv_act}
\alias{Ddv_act}
\title{Change in Activation Energy (Internal)}
\usage{
Ddv_act(wt, mut, ideal = wt, pdb_site_active = NA)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}

\item{ideal}{Reference protein for ideal active site conformation}

\item{pdb_site_active}{Vector of PDB residue numbers defining the active site}
}
\value{
a scalar energy difference (NA if pdb_site_active is NA)
}
\description{
Calculates the change in internal energy contribution to activation between
wild-type and mutant.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Ddg_ent_act}()},
\code{\link{Dg_ent}()},
\code{\link{Dmsfi}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
