% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{delta_energy}
\alias{delta_energy}
\alias{Dv_min}
\alias{Dg_ent}
\alias{Ddv_act}
\alias{Ddg_ent_act}
\title{Calculate energy differences between two proteins}
\usage{
Dv_min(wt, mut)

Dg_ent(wt, mut, beta = beta_boltzmann())

Ddv_act(wt, mut, ideal = wt, pdb_site_active = NA)

Ddg_ent_act(wt, mut, ideal = wt, pdb_site_active = NA, beta = beta_boltzmann())
}
\arguments{
\item{wt}{A protein object}

\item{mut}{A second protein}
}
\value{
A (scalar) energy difference between mutant and wild type.
}
\description{
Calculate energy differences between two proteins
}
\details{
`Dv_min` calculates the minimum-energy difference between \code{mut} and \code{wt}

`Dg_ent` calculates the entropic free energy difference between \code{mut} and \code{wt}

`Ddv_act` calculates the energy contribution to the change in activation energy between \code{mut} and \code{wt}

`Ddg_ent_act` calculates the entropy contribution to the change in activation energy between \code{mut} and \code{wt}
}
