% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_motion_by_site.R
\name{delta_motion_dmsfi}
\alias{delta_motion_dmsfi}
\title{Calculate site-dependent motion difference between two proteins}
\usage{
delta_motion_dmsfi(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object with \code{enm} defined}
}
\value{
A vector \code{(dmsf_i)} of size \code{nsites}, the MSF change at site i.
}
\description{
This version works only for wt and mut with no indels
}
