% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_motion_by_mode.R
\name{delta_motion_dmsfn}
\alias{delta_motion_dmsfn}
\title{Calculate mode-dependent motion difference between two proteins}
\usage{
delta_motion_dmsfn(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object with \code{enm} defined}
}
\value{
A vector \code{(dmsf_n)} of size \code{nmodes}, the MSF change in mode n.
}
\description{
This version works only for wt and mut with no indels.
Calculates fluctuations along normal modes of wt, independent of mode assignment issues.
}
