% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_structure_by_site.R
\name{delta_structure_dr2i}
\alias{delta_structure_dr2i}
\title{Calculate site-dependent structural difference between two proteins}
\usage{
delta_structure_dr2i(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} defined}

\item{mut}{A second protein object  with \code{xyz} defined}
}
\value{
A vector \code{(dr2_i)} of size \code{nsites}, the squared displacement at site i.
}
\description{
This version works only for wt and mut with no indels
}
