% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dg_ent}
\alias{Dg_ent}
\title{Change in Entropic Free Energy}
\usage{
Dg_ent(wt, mut, beta = beta_boltzmann())
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}

\item{beta}{Inverse temperature (1/kT). Default: beta_boltzmann()}
}
\value{
a scalar energy difference
}
\description{
Calculates the change in entropic free energy between wild-type and mutant.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Ddg_ent_act}()},
\code{\link{Ddv_act}()},
\code{\link{Dmsfi}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
