% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{dg_ent_act}
\alias{dg_ent_act}
\title{Activation free energy, entropic contribution}
\usage{
dg_ent_act(prot, ideal, pdb_site_active = NA, beta = beta_boltzmann())
}
\description{
Activation free energy, entropic contribution
}
\seealso{
Other enm_energy: 
\code{\link{dv_act}()},
\code{\link{g_ent}()},
\code{\link{v_min}()}
}
\concept{enm_energy}
