% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dmsfi}
\alias{Dmsfi}
\title{Calculate site-dependent motion difference between two proteins}
\usage{
Dmsfi(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{enm} defined}

\item{mut}{A second protein object with \code{enm} defined}
}
\value{
A vector \code{(dmsf_i)} of size \code{nsites}, the MSF change at site i.
}
\description{
This version works only for wt and mut with no indels
}
