% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dmsfn}
\alias{Dmsfn}
\title{Change in Mean-Square Fluctuation per Mode}
\usage{
Dmsfn(wt, mut)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\value{
a vector of size nmodes
}
\description{
Calculates the change in MSF projected onto each normal mode of the wild-type.
}
\details{
Note: Requires the `sclfenm` mutation model to be meaningful.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Dg_ent}()},
\code{\link{Dmsfi}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
