% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dr2n}
\alias{Dr2n}
\title{Squared Displacement per Mode}
\usage{
Dr2n(wt, mut)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\value{
a vector of size nmodes with squared displacements
}
\description{
Calculates the squared displacement projected onto each normal mode of the
wild-type.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Dg_ent}()},
\code{\link{Dmsfi}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
