% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dv_min}
\alias{Dv_min}
\title{Change in Minimum Energy}
\usage{
Dv_min(wt, mut)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\value{
a scalar energy difference
}
\description{
Calculates the change in minimum (stress) energy between wild-type and mutant.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Dg_ent}()},
\code{\link{Dmsfi}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()}
}
\concept{mutation-effect functions}
