% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{enm}
\alias{enm}
\title{Create an Elastic Network Model}
\usage{
enm(pdb, node, model, d_max)
}
\arguments{
\item{pdb}{pdb object obtained using bio3d::read.pdb}

\item{node}{parameter specifying how network nodes should be built: "sc" (side chains), "ca" (alpha carbons), or "cb" (beta carbons)}

\item{model}{parameter specifying model type: "anm", "ming_wall", "hnm", "hnm0", "pfanm", "reach"}

\item{d_max}{distance cutoff used to define enm contacts}
}
\value{
an object of class `prot`, which is a list `lst(param, node, graph, eij, kmat, nma)`
}
\description{
`enm` creates a `prot` object containing information on ENM structure, parameters, and normal modes
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5)
enm(pdb, node = "sc", model = "anm", d_max = 12.5)
enm(pdb, node = "cb", model = "anm", d_max = 12.0)
}
}
