% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{g_ent}
\alias{g_ent}
\title{Entropic Free Energy}
\usage{
g_ent(prot, beta = beta_boltzmann())
}
\arguments{
\item{prot}{A protein object created by [enm()]}

\item{beta}{Inverse temperature (1/kT). Default: beta_boltzmann()}
}
\value{
a scalar energy value
}
\description{
Calculates the entropic free energy contribution from vibrational modes.
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{dg_ent_act}()},
\code{\link{dv_act}()},
\code{\link{msfi}()},
\code{\link{msfn}()},
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
