% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{get_mlms}
\alias{get_mlms}
\title{Calculate MLMS site-dependent profile}
\usage{
get_mlms(prot, sdij_cut = 2)
}
\arguments{
\item{prot}{is a protein object obtained using enm()}

\item{sdij_cut}{An integer cutoff of sequence distance to include in calculation}
}
\value{
the profile of mean-local-mutational-stress (mlms) values
}
\description{
Calculates the Mean Local Mutational Stress (MLMS) profile using graph of prot object
}
\seealso{
Other site profiles: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{msfi}()},
\code{\link{wcn}()}
}
\concept{site profiles}
