% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{msfi}
\alias{msfi}
\title{Mean-Square Fluctuation per Site}
\usage{
msfi(prot)
}
\arguments{
\item{prot}{A protein object created by [enm()]}
}
\value{
a vector of size nsites with msf values for each site
}
\description{
Calculates the mean-square fluctuation of each site (diagonal of the
reduced covariance matrix).
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{dg_ent_act}()},
\code{\link{dv_act}()},
\code{\link{g_ent}()},
\code{\link{msfn}()},
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
