% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{msfn}
\alias{msfn}
\title{Mean-Square Fluctuation per Mode}
\usage{
msfn(prot)
}
\arguments{
\item{prot}{A protein object created by [enm()]}
}
\value{
a vector of size nmodes with msf values for each mode
}
\description{
Calculates the mean-square fluctuation contribution from each normal mode
(1/eigenvalue).
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{g_ent}()},
\code{\link{msfi}()},
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
