% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutenm-package.R
\docType{package}
\name{mutenm-package}
\alias{mutenm}
\alias{mutenm-package}
\title{mutenm: Mutate Elastic Network Models of Proteins}
\description{
Build elastic network models (ENMs) of proteins and calculate mutation-response matrices using the linear-force ENM (LFENM) perturbation model. Provides tools for analyzing how mutations propagate structural effects through proteins.
}
\details{
The \code{mutenm} package builds elastic network models (ENMs) of proteins
 and calculates mutation-response matrices using the linear-force ENM (LFENM)
 perturbation model.
}
