% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutenm.R
\name{mutenm}
\alias{mutenm}
\title{Mutate an Elastic Network Model}
\usage{
mutenm(
  wt,
  site_mut,
  mutation,
  mut_model = "lfenm",
  mut_dl_sigma = 0.3,
  mut_sd_min = 2,
  seed = 241956
)
}
\arguments{
\item{wt}{The protein \code{prot} to mutate}

\item{site_mut}{The site to mutate (not the pdb_site, but sequential)}

\item{mutation}{An integer (required). If 0, returns \code{wt} unchanged (with warning). Used with seed to generate reproducible random perturbations.}

\item{mut_model}{A string specifying mutational model ("lfenm" or "sclfenm")}

\item{mut_dl_sigma}{The standard deviation of a normal distribution from which edge-length perturbation is picked.}

\item{mut_sd_min}{An integer, only edges with \code{sdij >= mut_sd_min} are mutated}

\item{seed}{An integer, the seed for set.seed before picking perturbations}
}
\value{
A mutated protein object
}
\description{
Generates a single-point mutant of a protein ENM by perturbing contacts at a specified site.
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
wt <- set_enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5)
mut <- mutenm(wt, site_mut = 11, mutation = 1, mut_model = "lfenm", seed = 1024)
}

}
\concept{enm mutating functions}
