% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{v_min}
\alias{v_min}
\title{Calculate minimum energy of a given prot object}
\usage{
v_min(prot)
}
\arguments{
\item{prot}{is a prot object, wit a component graph tibble
where v0ij, kij, lij and the dij for the minimum conformation are found.}
}
\value{
a scalar: the energy at the minimum-enegy conformation
}
\description{
Calculate minimum energy of a given prot object
}
\seealso{
Other enm_energy: 
\code{\link{dg_ent_act}()},
\code{\link{dv_act}()},
\code{\link{g_ent}()}
}
\concept{enm_energy}
