% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{cn}
\alias{cn}
\title{Contact Number}
\usage{
cn(prot)
}
\arguments{
\item{prot}{A protein object created by [enm()]}
}
\value{
a vector of size nsites with cn values for each site
}
\description{
Calculates the Contact Number (CN) of each site - the number of neighbors
within the cutoff distance.
}
\seealso{
Other single-protein functions: 
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
