% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_energy.R
\name{delta_energy}
\alias{delta_energy}
\alias{Dv_min}
\alias{delta_energy_tds}
\alias{delta_energy_dvs}
\alias{delta_energy_act_dv}
\alias{delta_energy_act_tds}
\title{Calculate energy differences between a mutant and wild type}
\usage{
Dv_min(wt, mut)

delta_energy_tds(wt, mut, beta = beta_boltzmann())

delta_energy_dvs(wt, mut, ideal = wt)

delta_energy_act_dv(wt, mut, ideal = wt, pdb_site_active = NA)

delta_energy_act_tds(
  wt,
  mut,
  ideal = wt,
  pdb_site_active = NA,
  beta = beta_boltzmann()
)
}
\arguments{
\item{wt}{A protein object}

\item{mut}{A second protein}
}
\value{
A (scalar) energy difference between mutant and wild type.
}
\description{
Calculate energy differences between a mutant and wild type
}
\details{
`Dv_min` calculates the minimum-energy difference between \code{mut} and \code{wt}

`delta_energy_tds` calculates the entropic free energy difference between \code{mut} and \code{wt}

`delta_energy_dvs` calculates the ideal-conformation stress-energy difference between \code{mut} and \code{wt}

`delta_energy_act_dv` calculates the energy contribution to the change in activation energy between \code{mut} and \code{wt}

`delta_energy_act_tds` calculates the entropy contribution to the change in activation energy between \code{mut} and \code{wt}
}
