% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{dg_ent_act}
\alias{dg_ent_act}
\title{Activation Energy (Entropic)}
\usage{
dg_ent_act(prot, ideal, pdb_site_active = NA, beta = beta_boltzmann())
}
\arguments{
\item{prot}{A protein object created by [enm()]}

\item{ideal}{Reference protein for ideal active site conformation}

\item{pdb_site_active}{Vector of PDB residue numbers defining the active site}

\item{beta}{Inverse temperature (1/kT). Default: beta_boltzmann()}
}
\value{
a scalar energy value (NA if pdb_site_active is NA)
}
\description{
Calculates the entropic contribution to activation energy.
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{dv_act}()},
\code{\link{g_ent}()},
\code{\link{msfi}()},
\code{\link{msfn}()},
\code{\link{v_min}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
