% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dmsfi}
\alias{Dmsfi}
\title{Change in Mean-Square Fluctuation per Site}
\usage{
Dmsfi(wt, mut)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\value{
a vector of size nsites
}
\description{
Calculates the change in mean-square fluctuation at each site between
wild-type and mutant.
}
\details{
Note: Requires the `sclfenm` mutation model to be meaningful.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Ddg_ent_act}()},
\code{\link{Ddv_act}()},
\code{\link{Dg_ent}()},
\code{\link{Dmsfn}()},
\code{\link{Dr2i}()},
\code{\link{Dr2n}()},
\code{\link{Dv_min}()}
}
\concept{mutation-effect functions}
