% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dr2i}
\alias{Dr2i}
\title{Squared Displacement per Site}
\usage{
Dr2i(wt, mut)
}
\arguments{
\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\value{
a vector of size nsites with squared displacements
}
\description{
Calculates the squared displacement at each site between wild-type and mutant.
}
\seealso{
Other mutation-effect functions: 
\code{\link{Dr2n}()}
}
\concept{mutation-effect functions}
