% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_comparison.R
\name{Dr2n}
\alias{Dr2n}
\title{Calculate mode-dependent structural difference between two proteins}
\usage{
Dr2n(wt, mut)
}
\arguments{
\item{wt}{A protein object with \code{xyz} and \code{enm} defined}

\item{mut}{A second protein object with \code{xyz} defined}
}
\value{
A vector \code{(dr2_n)} of size \code{nmodes}, the squared displacement in mode n.
}
\description{
This version works only for wt and mut with no indels
}
