% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy_activation.R
\name{dv_act}
\alias{dv_act}
\title{Activation free energy, internal energy contribution}
\usage{
dv_act(prot, ideal, pdb_site_active = NA)
}
\description{
Activation free energy, internal energy contribution
}
\seealso{
Other enm_energy: 
\code{\link{dgact_tds}()},
\code{\link{g_ent}()},
\code{\link{v_min}()}
}
\concept{enm_energy}
