% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_energy.R
\name{g_ent}
\alias{g_ent}
\title{Calculate entropic total free energy of prot object}
\usage{
g_ent(prot, beta)
}
\arguments{
\item{prot}{is a prot object with known eigenvalues of enm model}

\item{beta}{is 1 / kT}
}
\description{
Calculate entropic total free energy of prot object
}
\seealso{
Other enm_energy: 
\code{\link{dg_ent_act}()},
\code{\link{dv_act}()},
\code{\link{v_min}()}
}
\concept{enm_energy}
