% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm_analysis.R
\name{msfi}
\alias{msfi}
\title{Calculate MSF site-dependent profile}
\usage{
msfi(prot)
}
\arguments{
\item{prot}{is a protein object obtained using enm()}
}
\value{
a vector of size nsites with msf values for each site
}
\description{
Calculates the mean-square-fluctuation of each site
}
\seealso{
Other site profiles: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{get_mlms}()},
\code{\link{wcn}()}
}
\concept{site profiles}
