% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutenm-package.R
\name{mutenm-params}
\alias{mutenm-params}
\title{Common parameters for mutenm functions}
\arguments{
\item{prot}{A protein object created by [enm()]}

\item{wt}{Wild-type protein object created by [enm()]}

\item{mut}{Mutant protein object created by [mutenm()]}
}
\description{
Common parameters for mutenm functions
}
\keyword{internal}
