% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{v_min}
\alias{v_min}
\title{Minimum Energy}
\usage{
v_min(prot)
}
\arguments{
\item{prot}{A protein object created by [enm()]}
}
\value{
a scalar energy value
}
\description{
Calculates the minimum (stress) energy of the protein at its current
conformation.
}
\seealso{
Other single-protein functions: 
\code{\link{cn}()},
\code{\link{wcn}()}
}
\concept{single-protein functions}
