% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_protein.R
\name{wcn}
\alias{wcn}
\title{Calculate WCN site-dependent profile}
\usage{
wcn(prot)
}
\arguments{
\item{prot}{is a protein object obtained using enm()}
}
\value{
a vector of size nsites with wcn values for each site
}
\description{
Calculates the Weighted Contact Number (WCN) of each site
}
\seealso{
Other site profiles: 
\code{\link{cn}()},
\code{\link{dactive}()},
\code{\link{msfi}()}
}
\concept{site profiles}
